# Conflicts:
#   backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/food/controller/FoodController.java

Former-commit-id: 04108f1e282fd4cf0427a46c2949009d285cd5f5
Former-commit-id: ff5d89b5b3c1e70de24bf78a080a0caa46661114

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/food/controller/FoodController.java
@@ -28,7 +28,7 @@ public class FoodController {
 	}
 
 	@GetMapping("/{longtitude}/{latitude}/{distance}/{sorted}")
-	public List<FoodResponseDto> getFoodFromLngLat(@PathVariable("longtitude")Double longitude, @PathVariable("latitude")Double latitude, @PathVariable("distance")Long distance, @PathVariable("sorted")String sorted){
+	public List<FoodResponseDto> getFoodFromLngLat(@PathVariable("longtitude")Double longitude, @PathVariable("latitude")Double latitude, @PathVariable("distance")int distance, @PathVariable("sorted")String sorted){
 		AttractionCoordinateRequestDto attractionCoordinateRequestDto = new AttractionCoordinateRequestDto(latitude, longitude, distance, sorted);
 		return foodServiceImpl.getFoodFromLngLat(attractionCoordinateRequestDto);
 	}
